+++ database/database_virtual_accounts.go
@@ -6,6 +6,30 @@ import (
 	"github.com/ndv6/tsaving/models"
 )
 
+func CheckBalance(target string, accNumber string, amount int, db *sql.DB) (status bool) {
+	if target == "MAIN" {
+		sourceBalance, err := GetBalanceAcc(accNumber, db)
+		if err != nil {
+			return
+		}
+		if sourceBalance.AccountBalance < amount || amount <= 0 {
+			return
+		}
+		status = true
+	}
+	if target == "VA" {
+		sourceBalance, err := GetBalanceVA(accNumber, db)
+		if err != nil {
+			return
+		}
+		if sourceBalance.VaBalance < amount || amount <= 0 {
+			return
+		}
+		status = true
+	}
+	return
+}
+
 func GetBalanceVA(vaNum string, db *sql.DB) (va models.VirtualAccounts, err error) {
 	var balanceFloat float32
 	err = db.QueryRow("SELECT va_balance FROM virtual_accounts WHERE va_num = ($1) ", vaNum).Scan(&balanceFloat)
